funcoes principai

+++ src/main/java/com/murilo_dev/system_pedidos/DTO/DadosPedidosDto.java
@@ -1,11 +1,12 @@
 package com.murilo_dev.system_pedidos.DTO;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.LocalDateTime;
 
-public record DadosPedidosDto(Long id,
+public record DadosPedidosDto(Integer id,
                               String nome_user,
                               String item_cardapio,
-                              double valor,
+                              BigDecimal valor,
                               LocalDateTime data_hora_pedido) {
 }
